+++ test/cases/validate_format_matcher_test.rb
@@ -5,6 +5,12 @@ class ValidateFormatMatcherTest < MiniTest::Unit::TestCase
     assert_raises(ArgumentError) { assert_must validate_format_of(:email), User}
   end
 
+  test 'raises if to_allow and to_not_allow are used' do
+    assert_raises RuntimeError do
+      assert_must validate_format_of(:email).to_allow('').to_not_allow(''), User
+    end
+  end
+
   test 'must validate valid format of an attribute' do
     assert_must validate_format_of(:email).to_allow('foo@bar.com'), User
   end